Abstract core functionality of MonoDependency and MonoDependent to enable testing.
