+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/BaseMonoDependency.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 05ec2c92bf19840499f8ad1fc41a2aff
