+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/BaseMonoDependent.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b5c6e180cbf28d84fbc2bd3c7a971eaf
