+++ Unity/com.chopsticks.dependencies/Assets/Scripts/Runtime/MonoDependency.cs
@@ -1,53 +1,11 @@
 ﻿using Chopsticks.Dependencies.Contained;
+using Chopsticks.Dependencies.Containers;
+using Chopsticks.Dependencies.Factories;
 
 namespace Chopsticks.Dependencies
 {
     /// <inheritdoc cref="IMonoDependency"/>
-    public abstract class MonoDependency : MonoDependent, IMonoDependency
-    {
-        ///<inheritdoc/>
-        public override void OnEnable()
-        {
-            base.OnEnable();
-            OnRegistration();
-        }
-
-        ///<inheritdoc/>
-        public virtual void OnDisable() => Deregister();
-
-        ///<inheritdoc/>
-        public override void OnTransformParentChanged()
-        {
-            if (!enabled)
-                return;
-
-            if (!GetUpdatedContainer(out var updatedContainer))
-                return;
-
-            Deregister();
-            Container = updatedContainer;
-            OnRegistration();
-
-            OnContainerChanged();
-        }
-
-
-        /// <summary>
-        /// Performs registration of this dependency for each of its contracts 
-        /// using <see cref="RegisterAs{T}"/>.
-        /// </summary>
-        protected abstract void OnRegistration();
-
-        protected void RegisterAs<T>()
-        {
-            // TODO :: Uses interface extension method to register.
-            // TODO :: Tracks registrations.
-        }
-
-
-        private void Deregister()
-        {
-            // TODO :: Deregisters all tracked registrations.
-        }
-    }
+    public abstract class MonoDependency : BaseMonoDependency<DependencyContainer,
+            UnityContainerService<DependencyContainer, DefaultDependencyContainerFactory,
+            DependencyContainerDefinition>>, IMonoDependency { }
 }
